Rename from XMLOutputTester to SVGBuilder.
